requirements.txt
amqp==2.3.2
Babel==2.6.0
billiard==3.5.0.4
bootstrap-admin==0.3.9
boto3==1.7.67
botocore==1.10.67
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.7
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-debug-toolbar==1.9.1
django-extensions==2.1.0
django-hosts==3.0
django-localflavor==2.0
django-phonenumber-field==2.0.0
django-postgrespool==0.3.1
django-sendgrid-v5==0.6.893
django-sortedm2m==1.5.0
django-storages==1.6.6
django-tastypie==0.14.1
django-timezone-field==2.1
django-url-or-relative-url-field==0.1.1
docutils==0.14
future==0.16.0
gevent==1.3.5
greenlet==0.4.13
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
kombu==4.2.1
lxml==4.2.3
newrelic==4.0.0.99
oauthlib==2.1.0
pdfminer.six==20170720
pdfplumber==0.5.10
phonenumberslite==8.9.10
Pillow==5.2.0
psycopg2-binary==2.7.5
pycryptodome==3.6.4
python-dateutil==2.7.3
python-http-client==3.1.0
python-mimeparse==1.6.0
python-postmark==0.5.0
python-twitter==3.4.2
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
sendgrid==5.4.1
six==1.11.0
SQLAlchemy==1.2.10
sqlparse==0.2.4
urllib3==1.23
vine==1.1.4
Wand==0.4.4
whitenoise==3.3.1
